qr-pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qr-pipeline"
version = "0.1.0"
description = "Query generation + (query, pos, neg) dataset builder for reranker fine-tuning."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Your Name"}]
dependencies = [
  "pyyaml>=6.0.1",
  "orjson>=3.10.0",
  "tqdm>=4.66.0",
  "regex>=2024.0.0",
  "python-dateutil>=2.9.0.post0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6.0",
]

# Optional (when you add local LLM inference):
# llm = [
#   "torch>=2.2.0",
#   "transformers>=4.41.0",
#   "accelerate>=0.30.0",
# ]

[project.scripts]
qr = "qr_pipeline.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
